refactor: - group types in their own files - group decorators in their own files - update store methods to be arrow functions
