+++ src/types/permission-types.ts
@@ -1,0 +1,1 @@
+export type Permission = "r" | "w" | "rw" | "none";
