+++ src/types/store-types.ts
@@ -1,0 +1,21 @@
+import { Store } from "../store";
+import { JSONArray, JSONObject, JSONPrimitive } from "./json-types";
+import { Permission } from "./permission-types";
+
+export type StoreResult = Store | JSONPrimitive | undefined;
+
+export type StoreValue =
+  | JSONObject
+  | JSONArray
+  | StoreResult
+  | (() => StoreResult);
+
+export interface IStore {
+  defaultPolicy: Permission;
+  allowedToRead(key: string): boolean;
+  allowedToWrite(key: string): boolean;
+  read(path: string): StoreResult;
+  write(path: string, value: StoreValue): StoreValue;
+  writeEntries(entries: JSONObject): void;
+  entries(): JSONObject;
+}
