+++ src/userStore.ts
@@ -1,4 +1,5 @@
-import { Restrict, Store } from "./store";
+import { Store } from "./store";
+import { Restrict } from "./utils/decorators";
 
 export class UserStore extends Store {
   @Restrict("rw")
