+++ src/utils/consts.ts
@@ -1,0 +1,2 @@
+export const INTERNAL_METHODS_KEY = Symbol("store-internal-methods-key");
+export const PERMISSIONS_KEY = Symbol("store-permissions-key");
